@@ -1,47 +1,38 @@
-import React from 'react';
+// src/App.jsx
+import React, { useState } from 'react';
 import './App.scss';
 import Header from './Components/Header';
 import Footer from './Components/Footer';
 import Form from './Components/Form';
 import Results from './Components/Results';
-import axios from 'axios'; // Import axios for making HTTP requests
+import axios from 'axios';
 
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      data: null,
-      requestParams: {},
-      error: null, // Add error state to handle errors
-    };
-  }
+const App = () => {
+  const [data, setData] = useState(null);
+  const [requestParams, setRequestParams] = useState({});
+  const [error, setError] = useState(null);
 
-  // Function to make API call
-  callApi = async (requestParams) => {
+  const callApi = async (requestParams) => {
     try {
-      // Make HTTP request using axios
       const response = await axios.get(requestParams.url);
-      // Update state with fetched data and request parameters
-      this.setState({ data: response.data, requestParams, error: null });
+      setData(response.data);
+      setRequestParams(requestParams);
+      setError(null);
     } catch (error) {
-      // Handle errors by setting error state
-      this.setState({ error: error.message });
+      setError(error.message);
     }
-  }
+  };
 
-  render() {
-    return (
-      <React.Fragment>
-        <Header />
-        <div>Request Method: {this.state.requestParams.method}</div>
-        <div>URL: {this.state.requestParams.url}</div>
-        <Form handleApiCall={this.callApi} />
-        {/* Pass data and error states as props to Results */}
-        <Results data={this.state.data} error={this.state.error} />
-        <Footer />
-      </React.Fragment>
-    );
-  }
-}
+  return (
+    <React.Fragment>
+      <Header />
+      <div>Request Method: {requestParams.method}</div>
+      <div>URL: {requestParams.url}</div>
+      <Form handleApiCall={callApi} />
+      <Results data={data} error={error} />
+      <Footer />
+    </React.Fragment>
+  );
+};
 
 export default App;
