@@ -1,6 +1,6 @@
+// src/main.jsx
 import React from 'react';
-import ReactDOM from 'react-dom/client';
-
+import ReactDOM from 'react-dom/client'; // Corrected import statement
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
