@@ -1,7 +1,12 @@
-import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite'
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  test: {
+    global: true,
+    environment: 'jsdom',
+    setupFiles: ['./tests.config.js'], // Changed to an array
+  }
 })
